Added in some more folder structure and moved the socket.io code from the server file to their own.
